Refactor timing handling by replacing TimingEnum with TimingMode and updating related components
