+++ src/com/digero/maestro/view/ProjectFrame.java
@@ -146,7 +146,7 @@ public class ProjectFrame extends JFrame implements TableLayoutConstants, ICompi
 	private JButton resetTempoButton;
 	private JFormattedTextField keySignatureField;
 	private JFormattedTextField timeSignatureField;
-    private JComboBox<TimingEnum> timingCombo;
+    private JComboBox<TimingMode> timingCombo;
 
     private JCheckBox tempoOnlyFirstCheckBox;
 	private JComboBox<Chord.CalcDynamics> dynaCombo;
@@ -675,13 +675,15 @@ public class ProjectFrame extends JFrame implements TableLayoutConstants, ICompi
 			});
 		}
 
-        timingCombo = new JComboBox<>(TimingEnum.values());
+        timingCombo = new JComboBox<>(TimingMode.values());
 
         timingCombo.addActionListener(e -> {
-            TimingEnum enm = ((TimingEnum) Objects.requireNonNull(timingCombo.getSelectedItem()));
-            timingCombo.setToolTipText(enm.getTooltip());
+            TimingMode mode = (TimingMode) Objects.requireNonNull(timingCombo.getSelectedItem());
+            timingCombo.setToolTipText(mode.getTooltip());
 
-            enm.action(abcSong);
+            if (abcSong != null) {
+                abcSong.setTimings(mode.organic, mode.multistage, mode.mixTimings, mode.swing, mode.priority, mode.upgraded);
+            }
 
             refreshPreviewSequence(false);
         });
@@ -1889,7 +1891,7 @@ public class ProjectFrame extends JFrame implements TableLayoutConstants, ICompi
 				// one or more timing settings were change in abc song
 
 				// setting on model dont fire action listener
-				timingCombo.getModel().setSelectedItem(TimingEnum.getInstance(abcSong.isOrganic(), abcSong.isOrganic2(), abcSong.isMixTiming(), abcSong.isTripletTiming(), abcSong.isPriorityActive(), abcSong.isUpgraded()));
+				timingCombo.getModel().setSelectedItem(TimingMode.getInstance(abcSong.isOrganic(), abcSong.isOrganic2(), abcSong.isMixTiming(), abcSong.isTripletTiming(), abcSong.isPriorityActive(), abcSong.isUpgraded()));
 
 				updateButtons(false);
 				break;
@@ -2208,7 +2210,7 @@ public class ProjectFrame extends JFrame implements TableLayoutConstants, ICompi
 		tempoSpinner.setValue(MidiConstants.DEFAULT_TEMPO_BPM);
 		keySignatureField.setValue(KeySignature.C_MAJOR);
 		timeSignatureField.setValue(TimeSignature.FOUR_FOUR);
-        timingCombo.getModel().setSelectedItem(TimingEnum.MIX);
+        timingCombo.getModel().setSelectedItem(TimingMode.MIX);
         dynaCombo.setSelectedItem(AbcSong.dynamicsMethodDefault);
         tempoOnlyFirstCheckBox.setSelected(false);
 		midiBarLabel.setBarNumberCache(null);
@@ -2317,7 +2319,7 @@ public class ProjectFrame extends JFrame implements TableLayoutConstants, ICompi
 			setMeter(abcSong.getTimeSignature());
 
             // setting on model dont fire action listener
-            timingCombo.getModel().setSelectedItem(TimingEnum.getInstance(abcSong.isOrganic(),abcSong.isOrganic2(),abcSong.isMixTiming(),abcSong.isTripletTiming(),abcSong.isPriorityActive(), abcSong.isUpgraded()));
+            timingCombo.getModel().setSelectedItem(TimingMode.getInstance(abcSong.isOrganic(),abcSong.isOrganic2(),abcSong.isMixTiming(),abcSong.isTripletTiming(),abcSong.isPriorityActive(), abcSong.isUpgraded()));
 
             tempoOnlyFirstCheckBox.setSelected(abcSong.isUsingOldTempos());
 
@@ -3631,96 +3633,6 @@ public class ProjectFrame extends JFrame implements TableLayoutConstants, ICompi
                     log.log(Level.WARNING, "Failed to connect to github for version check", e);
                 }
             });
-        }
-    }
-
-    public enum TimingEnum {
-        ORGANIC_MULTISTAGE2 (UIText.get("maestro.timing.organic.multi.stage.2"),true, true, false,false,false,"Organic Multistage 2", true),
-        ORGANIC_MULTISTAGE (UIText.get("maestro.timing.organic.multi.stage"),true, true, false,false,false,"Organic Multistage", false),
-        ORGANIC_SINGLESTAGE (UIText.get("maestro.timing.organic.single.stage"), true, false, false,false,false,"Organic Singlestage", false),
-        MIX (UIText.get("maestro.timing.mix.timings"), false, false, true,false,false,"Mix Timings", false),
-        MIX_SWING (UIText.get("maestro.timing.mix.timings.swing"), false, false, true,true,false,"Mix Timings Swing/Triplet", false),
-        MIX_PRIO (UIText.get("maestro.timing.mix.timings.combine.priorities"), false, false, true,false,true,"Mix Timings Combine Priorities", false),
-        MIX_SWING_PRIO (UIText.get("maestro.timing.mix.timings.swing.combine.priorities"), false, false, true,true,true,"Mix Timings Swing/Triplet Combine Priorities", false),
-        LEGACY (UIText.get("maestro.timing.legacy.timings"), false, false, false,false,false,"Legacy", false),
-        LEGACY_SWING (UIText.get("maestro.timing.legacy.timings.swing"), false, false, false,true,false,"Legacy Swing/Triplet", false),
-        ;
-
-        public final boolean organic;
-        public final boolean multistage;
-        public final boolean mixTimings;
-        public final boolean swing;
-        public final boolean priority;
-        public final String info;
-        public final String settingsString;// use this for settings prefs. And never change the strings.
-        public final boolean upgraded;
-
-        TimingEnum(String info, boolean organic, boolean multistage, boolean mixTimings, boolean swing, boolean priority, @NonNls String settings, boolean upgraded) {
-            this.info = info;
-            this.organic = organic;
-            this.multistage = multistage;
-            this.mixTimings = mixTimings;
-            this.swing = swing;
-            this.priority = priority;
-            this.settingsString = settings;
-            this.upgraded = upgraded;
-        }
-
-        public static TimingEnum getFromSettings(String defaultTiming) {
-            Objects.requireNonNull(defaultTiming);
-            for (TimingEnum timing : TimingEnum.values()) {
-                if (timing.settingsString.equals(defaultTiming)) {
-                    return timing;
-                }
-            }
-            return MIX;
-        }
-
-        public void action(@Nullable AbcSong abcSong) {
-            if (abcSong != null) {
-                abcSong.setTimings(organic, multistage, mixTimings, swing, priority, upgraded);
-            }
-        }
-
-        String getTooltip() {
-            return switch (this) {
-                case ORGANIC_MULTISTAGE2 -> UIText.get("maestro.tip.multi2");
-                case ORGANIC_MULTISTAGE -> UIText.get("maestro.tip.multi1");
-                case ORGANIC_SINGLESTAGE -> UIText.get("maestro.tip.single");
-                case LEGACY -> UIText.get("maestro.tip.legacy");
-                case LEGACY_SWING -> UIText.get("maestro.tip.legacy.swing");
-                case MIX_SWING_PRIO -> UIText.get("maestro.tip.mix.swing.prio");
-                case MIX -> UIText.get("maestro.tip.mix");
-                case MIX_SWING -> UIText.get("maestro.tip.mix.swing");
-                case MIX_PRIO -> UIText.get("maestro.tip.mix.prio");
-                default -> null;
-            };
-        }
-
-        static TimingEnum getInstance(boolean organic, boolean multistage, boolean mixTimings, boolean swing, boolean priority, boolean upgraded) {
-            if (organic) {
-                if (multistage) {
-                    if (upgraded) return ORGANIC_MULTISTAGE2;
-                    return ORGANIC_MULTISTAGE;
-                }
-                else return ORGANIC_SINGLESTAGE;
-            } else if (mixTimings) {
-                if (swing) {
-                    if (priority) return MIX_SWING_PRIO;
-                    else return MIX_SWING;
-                } else {
-                    if (priority) return MIX_PRIO;
-                    else return MIX;
-                }
-            } else {
-                if (swing) return LEGACY_SWING;
-                else return LEGACY;
-            }
-        }
-
-        @Override
-        public String toString() {
-            return info;
         }
     }
 }
